Distinguish between activation failure and attempted reactivation.
Add a boolean field , `activated` and corresponding migration
to `RegistrationProfile`. Use this field when the user tries
to activate their account to distinguish between:

- account activation failure due to non-existant activation key
- account activation failure due to an expired activation key
- account activation failure due to the user trying to activate
  more than once

By distinguising between these scenarios, it is possible to
provide the user with better information during account activation.
Specifically, if the user is active and tries to reactivate,
clearly indicate their account is active and able to login.

Update documentation and unit tests to account for additional
model field and activation logic.

Add both a schema and data migration. Data migration checks
if `activation_key` is the `ACTIVATED` string and conditionally
sets the new `activated` field.